requirements.txt
numpy
python-bitcoinlib
cryptography
pem
pycryptodomex
gmpy2
pyyaml
progressbar2
aiohttp
